refactor: consolidate cache subsystem into a single module
Collapse the 5 overlapping cache files (3788 LOC) into one consolidated
implementation in src/core/bidirectional-cache.ts plus the absorbing
smart-normalizer.ts. Mechanical only — no cache-logic changes.

Inlined:
- src/core/enhanced-cache-key.ts        -> bidirectional-cache.ts (EnhancedCacheKeyManager + 4 types)
- src/core/tiered-cache.ts              -> bidirectional-cache.ts (TieredCache class)
- src/core/enhanced-cache-integration.ts -> bidirectional-cache.ts (EnhancedCacheIntegration singleton)
- src/core/context-aware-similarity.ts  -> smart-normalizer.ts    (ContextAwareSimilarity + types + singleton)

The three dynamic require('./context-aware-similarity.js') calls in
smart-normalizer.ts are replaced by direct references now that the
circular-dep workaround is no longer needed.

TestStep is consolidated into src/types/common.ts as the single
authority. The two duplicate definitions in enhanced-cache-key.ts and
test-scenario-cache.ts are removed; downstream importers either import
from common.ts directly or via the re-export from test-scenario-cache.ts.

Public API (src/index.ts) drops 9 internalized symbols:
EnhancedCacheKeyManager, CacheKeyComponents, EnhancedCacheKey,
StepsStructureAnalysis, URLPatternComponents, ContextAwareSimilarity,
SIMILARITY_THRESHOLDS, contextAwareSimilarity, SimilarityContext.
SmartNormalizer + types and BidirectionalCache + CacheMigrationManager
remain exported.

Callers updated to the new import paths: src/mcp/server.ts:11,
src/cli/cache-commands.ts:7, src/commands/cache-debug.ts (3 imports),
src/core/cache-migration.ts:2.

Tests:
- bidirectional-cache.test.js: only the require() path for TieredCache
  changes; the two pre-existing failing suites (Snapshot metrics,
  TieredCache wrapper) remain identical pre/post refactor — not a
  regression.
- Standalone tests that exercised the now-internalized symbols are
  removed: enhanced-cache-key.test.js, context-aware-similarity.test.js,
  performance-benchmarks.test.js, phase2-integration.test.js,
  cross-environment-matching.test.js.
- build-verification.js was a one-shot "Phase 2.3 complete" milestone
  marker hardcoded to deleted files and removed symbols; deleted.
- docs/CACHING.md updated to point at the consolidated locations.

Known follow-up: bidirectional-cache.ts and cache-migration.ts now form
a circular import (cache-migration imports EnhancedCacheKeyManager from
bidirectional-cache; bidirectional-cache imports CacheMigrationManager
from cache-migration). Works at runtime because all usages are inside
class bodies (deferred), but the indirection should be untangled when
the cache is redesigned in Epic 2. Documented inline at both import
sites.

Closes #38
