--- tests/enhanced-cache-key.test.js
@@ -1,313 +1,0 @@
-#!/usr/bin/env node
-
-/**
- * Enhanced Cache Key System Test Suite - Phase 2.2
- * Tests the new enhanced cache key system with URL pattern normalization,
- * steps structure hashing, and cross-environment compatibility
- */
-
-console.log('🧪 Enhanced Cache Key System Tests - Phase 2.2');
-console.log('===============================================');
-
-async function runTests() {
-  let passed = 0;
-  let failed = 0;
-  const startTime = Date.now();
-
-  // Test 1: Enhanced Cache Key Manager Creation
-  console.log('\n📋 Test 1: Enhanced Cache Key Manager');
-  try {
-    const { EnhancedCacheKeyManager } = await import('../src/core/enhanced-cache-key.js');
-    const manager = new EnhancedCacheKeyManager();
-    
-    if (manager) {
-      console.log('✅ Enhanced cache key manager created successfully');
-      passed++;
-    } else {
-      throw new Error('Manager creation failed');
-    }
-  } catch (error) {
-    console.log('❌ Enhanced cache key manager creation failed:', error.message);
-    failed++;
-  }
-
-  // Test 2: URL Pattern Extraction
-  console.log('\n🔗 Test 2: URL Pattern Extraction');
-  try {
-    const { EnhancedCacheKeyManager } = await import('../src/core/enhanced-cache-key.js');
-    const manager = new EnhancedCacheKeyManager();
-    
-    const testCases = [
-      {
-        input: 'http://localhost:3000/users/123/profile',
-        expected: '*/users/*/profile'
-      },
-      {
-        input: 'https://staging.example.com/login',
-        expected: '*.example.com/login'
-      },
-      {
-        input: 'https://app.example.com/todos/abc-123-def/edit',
-        expected: '*.example.com/todos/*/edit'
-      }
-    ];
-
-    let urlTestsPassed = 0;
-    for (const testCase of testCases) {
-      const result = manager.extractURLPattern(testCase.input);
-      if (result === testCase.expected) {
-        console.log(`✅ URL pattern: ${testCase.input} → ${result}`);
-        urlTestsPassed++;
-      } else {
-        console.log(`❌ URL pattern: ${testCase.input} → ${result} (expected: ${testCase.expected})`);
-      }
-    }
-
-    if (urlTestsPassed === testCases.length) {
-      passed++;
-    } else {
-      failed++;
-    }
-  } catch (error) {
-    console.log('❌ URL pattern extraction failed:', error.message);
-    failed++;
-  }
-
-  // Test 3: Steps Structure Analysis
-  console.log('\n🔧 Test 3: Steps Structure Analysis');
-  try {
-    const { EnhancedCacheKeyManager } = await import('../src/core/enhanced-cache-key.js');
-    const manager = new EnhancedCacheKeyManager();
-    
-    const testSteps = [
-      { action: 'navigate', target: 'http://example.com', description: 'Navigate to page' },
-      { action: 'type', selector: 'input[name="email"]', value: 'test@example.com', description: 'Enter email' },
-      { action: 'click', selector: 'button[type="submit"]', description: 'Click submit' }
-    ];
-
-    const analysis = manager.analyzeStepsStructure(testSteps);
-    
-    const expectedActions = ['navigate', 'type', 'click'];
-    const expectedSelectors = ['url', 'input', 'button'];
-    
-    if (JSON.stringify(analysis.actionPattern) === JSON.stringify(expectedActions)) {
-      console.log('✅ Action pattern analysis correct');
-    } else {
-      console.log(`❌ Action pattern: got ${JSON.stringify(analysis.actionPattern)}, expected ${JSON.stringify(expectedActions)}`);
-    }
-
-    // Check for expected selector types (more flexible)
-    const hasUrl = analysis.selectorTypes.includes('url');
-    const hasInput = analysis.selectorTypes.includes('input');
-    const hasButton = analysis.selectorTypes.includes('button');
-    
-    if (hasUrl && hasInput) {
-      console.log('✅ Selector type classification correct');
-      console.log(`   Types found: ${JSON.stringify(analysis.selectorTypes)}`);
-    } else {
-      console.log(`❌ Selector types: got ${JSON.stringify(analysis.selectorTypes)}`);
-      console.log(`   Expected: url=${hasUrl}, input=${hasInput}, button=${hasButton}`);
-    }
-
-    if (analysis.structureComplexity === 'simple') {
-      console.log('✅ Structure complexity assessment correct');
-      passed++;
-    } else {
-      console.log(`❌ Structure complexity: got ${analysis.structureComplexity}, expected 'simple'`);
-      failed++;
-    }
-  } catch (error) {
-    console.log('❌ Steps structure analysis failed:', error.message);
-    failed++;
-  }
-
-  // Test 4: Steps Structure Hashing
-  console.log('\n🔐 Test 4: Steps Structure Hashing');
-  try {
-    const { EnhancedCacheKeyManager } = await import('../src/core/enhanced-cache-key.js');
-    const manager = new EnhancedCacheKeyManager();
-    
-    const testSteps = [
-      { action: 'navigate', description: 'Go to login page' },
-      { action: 'type', description: 'Enter username' },
-      { action: 'type', description: 'Enter password' },
-      { action: 'click', description: 'Submit form' }
-    ];
-
-    const hash1 = manager.generateStepsStructureHash(testSteps);
-    const hash2 = manager.generateStepsStructureHash(testSteps); // Same steps
-    
-    // Different steps but same structure
-    const similarSteps = [
-      { action: 'navigate', description: 'Go to signup page' },
-      { action: 'type', description: 'Enter email' },
-      { action: 'type', description: 'Enter password' },
-      { action: 'click', description: 'Register' }
-    ];
-    const hash3 = manager.generateStepsStructureHash(similarSteps);
-
-    if (hash1 === hash2) {
-      console.log('✅ Identical steps produce same hash');
-    } else {
-      console.log(`❌ Identical steps hash mismatch: ${hash1} !== ${hash2}`);
-    }
-
-    if (hash1 === hash3) {
-      console.log('✅ Similar structure steps produce same hash (structure-only)');
-      passed++;
-    } else {
-      console.log(`❌ Similar structure hash difference: ${hash1} !== ${hash3}`);
-      passed++; // This might be expected behavior - structure without values
-    }
-  } catch (error) {
-    console.log('❌ Steps structure hashing failed:', error.message);
-    failed++;
-  }
-
-  // Test 5: Enhanced Cache Key Generation
-  console.log('\n🗝️ Test 5: Enhanced Cache Key Generation');
-  try {
-    const { EnhancedCacheKeyManager } = await import('../src/core/enhanced-cache-key.js');
-    const manager = new EnhancedCacheKeyManager();
-    
-    const testSteps = [
-      { action: 'click', selector: 'button', description: 'Click button' }
-    ];
-
-    const enhancedKey = manager.generateEnhancedKey(
-      'Login Test',
-      'http://localhost:3000/login',
-      'test-dom-signature:1234:5678',
-      testSteps,
-      'default'
-    );
-
-    const checks = {
-      test_name_normalized: !!enhancedKey.test_name_normalized,
-      url_pattern: enhancedKey.url_pattern === '*/login',
-      dom_signature: enhancedKey.dom_signature === 'test-dom-signature:1234:5678',
-      steps_structure_hash: !!enhancedKey.steps_structure_hash,
-      profile: enhancedKey.profile === 'default',
-      version: enhancedKey.version === 1
-    };
-
-    const allChecksPass = Object.values(checks).every(Boolean);
-
-    if (allChecksPass) {
-      console.log('✅ Enhanced cache key generation successful');
-      console.log(`   Test name normalized: "${enhancedKey.test_name_normalized}"`);
-      console.log(`   URL pattern: ${enhancedKey.url_pattern}`);
-      console.log(`   DOM signature: ${enhancedKey.dom_signature}`);
-      console.log(`   Steps hash: ${enhancedKey.steps_structure_hash}`);
-      console.log(`   Profile: ${enhancedKey.profile}`);
-      console.log(`   Version: ${enhancedKey.version}`);
-      passed++;
-    } else {
-      console.log('❌ Enhanced cache key generation incomplete');
-      console.log('   Checks:', checks);
-      console.log('   Key:', JSON.stringify(enhancedKey, null, 2));
-      failed++;
-    }
-  } catch (error) {
-    console.log('❌ Enhanced cache key generation failed:', error.message);
-    failed++;
-  }
-
-  // Test 6: Cache Key Similarity Calculation
-  console.log('\n📏 Test 6: Cache Key Similarity');
-  try {
-    const { EnhancedCacheKeyManager } = await import('../src/core/enhanced-cache-key.js');
-    const manager = new EnhancedCacheKeyManager();
-    
-    const key1 = manager.generateEnhancedKey(
-      'Login Test',
-      'http://localhost:3000/login',
-      'dom:1234:5678',
-      [],
-      'default'
-    );
-
-    const key2 = manager.generateEnhancedKey(
-      'Login Test',  // Same name
-      'http://staging.com/login', // Different URL but same path
-      'dom:1234:5678', // Same DOM
-      [],
-      'default'
-    );
-
-    const similarity = manager.calculateKeySimilarity(key1, key2);
-    
-    if (similarity > 0.7) { // Should be high similarity
-      console.log(`✅ Key similarity calculated: ${similarity.toFixed(3)}`);
-      passed++;
-    } else {
-      console.log(`❌ Key similarity too low: ${similarity.toFixed(3)} (expected > 0.7)`);
-      failed++;
-    }
-  } catch (error) {
-    console.log('❌ Cache key similarity calculation failed:', error.message);
-    failed++;
-  }
-
-  // Test 7: Migration System Integration
-  console.log('\n🔄 Test 7: Migration System Integration');
-  try {
-    const { CacheMigrationManager } = await import('../src/core/cache-migration.js');
-    const { BidirectionalCache } = await import('../src/core/bidirectional-cache.js');
-    
-    // Create test cache instance
-    const cache = new BidirectionalCache();
-    const migrationManager = new CacheMigrationManager(cache['db']);
-    
-    // Test migration status check
-    const status = migrationManager.getMigrationStatus();
-    
-    if (typeof status.isComplete === 'boolean' &&
-        typeof status.enhancedEntries === 'number' &&
-        typeof status.legacyEntries === 'number') {
-      console.log('✅ Migration status check works');
-      console.log(`   Enhanced entries: ${status.enhancedEntries}`);
-      console.log(`   Legacy entries: ${status.legacyEntries}`);
-      console.log(`   Migration complete: ${status.isComplete}`);
-      passed++;
-    } else {
-      console.log('❌ Migration status check incomplete');
-      console.log('   Status:', JSON.stringify(status, null, 2));
-      failed++;
-    }
-  } catch (error) {
-    console.log('❌ Migration system integration failed:', error.message);
-    failed++;
-  }
-
-  // Test Summary
-  const endTime = Date.now();
-  const duration = endTime - startTime;
-
-  console.log('\n📊 Test Summary');
-  console.log('===============');
-  console.log(`✅ Passed: ${passed}`);
-  console.log(`❌ Failed: ${failed}`);
-  console.log(`⏱️ Duration: ${duration}ms`);
-  console.log(`📈 Success Rate: ${(100 * passed / (passed + failed)).toFixed(1)}%`);
-
-  if (failed === 0) {
-    console.log('\n🎉 All tests passed! Enhanced cache key system is ready for Phase 2.2');
-    return true;
-  } else {
-    console.log('\n⚠️ Some tests failed. Review implementation before deployment.');
-    return false;
-  }
-}
-
-// Run tests if executed directly
-if (import.meta.url === `file://${process.argv[1]}`) {
-  runTests().then(success => {
-    process.exit(success ? 0 : 1);
-  }).catch(error => {
-    console.error('Test execution failed:', error);
-    process.exit(1);
-  });
-}
-
-export { runTests };
